refactor: domain 내 클래스 패키지 분리
