+++ src/main/java/nextstep/ladder/domain/dimension/Point.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.dimension;
 
 import java.util.HashMap;
 import java.util.Map;
