+++ src/main/java/nextstep/ladder/domain/generator/BooleanGenerator.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.generator;
 
 public interface BooleanGenerator {
