+++ src/main/java/nextstep/ladder/domain/generator/RandomBooleanGenerator.java
@@ -1,4 +1,6 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.generator;
+
+import nextstep.ladder.domain.generator.BooleanGenerator;
 
 import java.util.Random;
