+++ src/main/java/nextstep/ladder/domain/person/People.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.person;
 
 import java.util.List;
