+++ src/test/java/nextstep/ladder/domain/LineTest.java
@@ -2,6 +2,9 @@ package nextstep.ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import nextstep.ladder.domain.dimension.Line;
+import nextstep.ladder.domain.generator.BooleanGenerator;
+import nextstep.ladder.domain.generator.RandomBooleanGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
