+++ src/test/java/nextstep/ladder/domain/PersonTest.java
@@ -3,7 +3,7 @@ package nextstep.ladder.domain;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import org.junit.jupiter.api.Assertions;
+import nextstep.ladder.domain.person.Person;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
