+++ src/test/java/nextstep/ladder/domain/PointTest.java
@@ -2,6 +2,7 @@ package nextstep.ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import nextstep.ladder.domain.dimension.Point;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
